# students/k3341/Moskalets_Danila/Lr1/teamfinder-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TeamFinder API
  description: Платформа для поиска людей в команду для совместной работы над проектами
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /users:
    get:
      tags: [users]
      summary: Поиск и фильтрация пользователей
      parameters:
        - name: skill_ids
          in: query
          description: Фильтр по ID навыков (можно несколько)
          schema:
            type: array
            items:
              type: integer
        - name: proficiency_level
          in: query
          schema:
            $ref: '#/components/schemas/ProficiencyLevel'
        - name: search
          in: query
          description: Поиск по имени или bio
          schema:
            type: string
        - name: sort_by
          in: query
          description: Поле для сортировки
          schema:
            type: string
            enum: [created_at, first_name, last_name]
            default: created_at
        - name: sort_order
          in: query
          description: Направление сортировки
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: Номер страницы (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Количество записей на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список пользователей с пагинацией
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/{user_id}:
    get:
      tags: [users]
      summary: Получить профиль пользователя
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me:
    patch:
      tags: [users]
      summary: Обновить свой профиль
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/me/change-password:
    post:
      tags: [users]
      summary: Смена пароля текущего пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                  description: Текущий пароль
                new_password:
                  type: string
                  minLength: 8
                  description: Новый пароль (минимум 8 символов)
      responses:
        '200':
          description: Пароль успешно изменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password changed successfully
        '400':
          description: Неверный текущий пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/me/stats:
    get:
      tags: [users]
      summary: Получить статистику текущего пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Статистика пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/skills:
    get:
      tags: [users]
      summary: Получить навыки текущего пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список навыков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSkillResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [users]
      summary: Добавить навык текущему пользователю
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserSkillRequest'
      responses:
        '201':
          description: Навык добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSkillResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Навык уже добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/skills/{skill_id}:
    delete:
      tags: [users]
      summary: Удалить навык у текущего пользователя
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      responses:
        '204':
          description: Навык удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── SKILLS ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: integer
      description: ID пользователя
    SkillId:
      name: skill_id
      in: path
      required: true
      schema:
        type: integer
      description: ID навыка
...
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    UserRole:
      type: string
      enum: [ADMIN, USER]
      description: Роль пользователя
    ProficiencyLevel:
      type: string
      enum: [BEGINNER, INTERMEDIATE, ADVANCED, EXPERT]
      description: Уровень владения навыком
...
    Pagination:
      type: object
      required: [total, page, page_size, total_pages]
      properties:
        total:
          type: integer
          description: Общее количество записей
          example: 150
        page:
          type: integer
          description: Текущая страница (1-based)
          minimum: 1
          example: 2
        page_size:
          type: integer
          description: Размер страницы
          minimum: 1
          maximum: 100
          example: 20
        total_pages:
          type: integer
          description: Всего страниц
          example: 8

    # ── AUTH ──
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        role:
          $ref: '#/components/schemas/UserRole'
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@example.com
        avatar_url:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        bio:
          type: string
          nullable: true
          example: Passionate developer
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
    UserDetailResponse:
      allOf:
        - $ref: '#/components/schemas/UserResponse'
        - type: object
          properties:
            skills:
              type: array
              items:
                $ref: '#/components/schemas/UserSkillResponse'
    UserListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 64
          example: John
        last_name:
          type: string
          maxLength: 64
          example: Doe
        bio:
          type: string
          example: Updated bio
        avatar_url:
          type: string
          example: https://example.com/new-avatar.jpg
    UserStatsResponse:
      type: object
      properties:
        total_projects_owned:
          type: integer
          description: Количество созданных проектов
          example: 5
        total_tasks_assigned:
          type: integer
          description: Количество назначенных задач
          example: 12
        total_teams:
          type: integer
          description: Количество команд, в которых состоит пользователь
          example: 3
        total_completed_tasks:
          type: integer
          description: Количество выполненных задач
          example: 8

    # ── SKILLS ──
    SkillResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Python
        category:
          type: string
          example: Programming
...
    UserSkillResponse:
      type: object
      properties:
        skill:
          $ref: '#/components/schemas/SkillResponse'
        proficiency_level:
          $ref: '#/components/schemas/ProficiencyLevel'
    AddUserSkillRequest:
      type: object
      required: [skill_id, proficiency_level]
      properties:
        skill_id:
          type: integer
          example: 1
        proficiency_level:
          $ref: '#/components/schemas/ProficiencyLevel'

    # ── PROJECTS ──
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
              example: 404
            message:
              type: string
              example: Resource not found
            details:
              type: object
              description: Дополнительные сведения об ошибке
              additionalProperties: true
              nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: integer
              example: 422
            message:
              type: string
              example: Validation error
            details:
              type: array
              items:
                type: object
                properties:
                  loc:
                    type: array
                    items:
                      type: string
                    example: ["body", "email"]
                  msg:
                    type: string
                    example: field required
                  type:
                    type: string
                    example: missing